+++ src/OpenTelemetry.Api/Metrics/CounterMetric.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 // </copyright>
 
+using System;
 using System.Collections.Generic;
 using OpenTelemetry.Trace;
 
@@ -23,6 +24,7 @@ namespace OpenTelemetry.Metrics
     /// Counter instrument.
     /// </summary>
     /// <typeparam name="T">The type of counter. Only long and double are supported now.</typeparam>
+    [Obsolete("Metrics API/SDK is not recommended for production. See https://github.com/open-telemetry/opentelemetry-dotnet/issues/1501 for more information on metrics support.")]
     public abstract class CounterMetric<T>
         where T : struct
     {