+++ src/OpenTelemetry/Metrics/Aggregators/Aggregator.cs
@@ -24,6 +24,7 @@ namespace OpenTelemetry.Metrics.Aggregators
     /// Aggregator base class.
     /// </summary>
     /// <typeparam name="T">The type.</typeparam>
+    [Obsolete("Metrics API/SDK is not recommended for production. See https://github.com/open-telemetry/opentelemetry-dotnet/issues/1501 for more information on metrics support.")]
     public abstract class Aggregator<T>
         where T : struct
     {